{"schema":"decodex/commit/1","summary":"Split account panel and run strip modules","authority":"manual"}
