+++ apps/decodex-app/Sources/DecodexApp/AccountPanelView.swift
@@ -1,17 +1,16 @@
 import AppKit
-import Combine
 import Foundation
 import SwiftUI
 
 struct AccountPanelView: View {
 	@ObservedObject var store: AccountStore
 	@ObservedObject var loginWindowState: LoginWindowState
-	@Environment(\.colorScheme) private var colorScheme
-	@State private var accountScrollOffset: CGFloat = 0
-	@State private var armedLogoutAccountID: String?
-	@State private var deletingLogoutAccountID: String?
-	@State private var logoutErrorMessage: String?
-	@AppStorage("decodex.operator.accountPrivacy") private var accountPrivacy = AccountPrivacy.hiddenValue
+	@Environment(\.colorScheme) var colorScheme
+	@State var accountScrollOffset: CGFloat = 0
+	@State var armedLogoutAccountID: String?
+	@State var deletingLogoutAccountID: String?
+	@State var logoutErrorMessage: String?
+	@AppStorage("decodex.operator.accountPrivacy") var accountPrivacy = AccountPrivacy.hiddenValue
 
 	var body: some View {
 		Group {
@@ -76,151 +75,6 @@ struct AccountPanelView: View {
 		.controlSize(.small)
 		.symbolRenderingMode(.hierarchical)
 		.animation(PanelMotion.panelLayout, value: panelAnimationKey)
-	}
-
-	private var header: some View {
-		HStack(alignment: .center, spacing: 8) {
-			Image(nsImage: AppAssets.statusBarIcon)
-				.resizable()
-				.renderingMode(.template)
-				.scaledToFit()
-				.foregroundStyle(PanelPalette.actionBlue(colorScheme))
-				.frame(width: 20, height: 20)
-				.frame(width: 28, height: 28)
-
-			VStack(alignment: .leading, spacing: 2) {
-				Text("Decodex")
-					.font(PanelFont.headerTitle)
-					.foregroundStyle(PanelPalette.primaryText(colorScheme))
-				Text(headerSubtitle)
-					.font(PanelFont.headerSubtitle)
-					.foregroundStyle(PanelPalette.secondaryText(colorScheme))
-					.lineLimit(1)
-					.minimumScaleFactor(0.9)
-			}
-			.layoutPriority(1)
-
-			Spacer(minLength: 4)
-
-			HStack(spacing: 5) {
-				PanelIconButtonView(
-					symbol: emailsHidden ? "eye.slash" : "eye",
-					tint: PanelPalette.secondaryText(colorScheme),
-					isActive: false,
-					action: {
-						withAnimation(PanelMotion.inlineLayout) {
-							accountPrivacy = emailsHidden ? AccountPrivacy.visibleValue : AccountPrivacy.hiddenValue
-						}
-					},
-					help: emailsHidden ? "Show account emails" : "Hide account emails"
-				)
-
-				PanelIconButtonView(
-					symbol: store.fastModeEnabled ? "bolt.fill" : "bolt",
-					tint: PanelPalette.fastModeAccent(colorScheme),
-					isActive: store.fastModeEnabled,
-					isDisabled: store.isSettingFastMode,
-					size: 25,
-					action: {
-						Task {
-							await store.setFastMode(store.fastModeEnabled == false)
-						}
-					},
-					help: store.fastModeEnabled ? "Turn fast mode off" : "Turn fast mode on"
-				)
-
-				PanelIconButtonView(
-					symbol: "safari",
-					tint: PanelPalette.actionBlue(colorScheme),
-					isActive: false,
-					action: {
-						Task {
-							await store.openWebUI()
-						}
-					},
-					help: "Open Decodex WebUI"
-				)
-
-				PanelIconButtonView(
-					symbol: "plus",
-					tint: PanelPalette.actionBlue(colorScheme),
-					isActive: false,
-					isPrimary: true,
-					size: 25,
-					action: {
-						presentLogin(.newAccount)
-					},
-					help: "Add login"
-				)
-			}
-		}
-		.animation(PanelMotion.state, value: hasFixedSelection)
-	}
-
-	private var accountSummary: some View {
-		HStack(alignment: .firstTextBaseline, spacing: 7) {
-			SummaryTileView(
-				title: "Codex",
-				value: codexAuthLabel,
-				symbol: "terminal",
-				tint: PanelPalette.codexAccent(colorScheme)
-			)
-
-			Rectangle()
-				.fill(PanelPalette.separator(colorScheme))
-				.frame(width: 0.5, height: 16)
-				.alignmentGuide(.firstTextBaseline) { dimensions in
-					dimensions[VerticalAlignment.center] + 4
-				}
-
-			SummaryTileView(
-				title: "Runs",
-				value: decodexModeLabel,
-				symbol: "arrow.triangle.branch",
-				tint: hasFixedSelection ? PanelPalette.actionBlue(colorScheme) : PanelPalette.secondaryText(colorScheme)
-			)
-		}
-		.padding(.horizontal, 3)
-		.padding(.top, 1)
-		.padding(.bottom, 4)
-		.overlay(alignment: .bottom) {
-			Rectangle()
-				.fill(PanelPalette.separator(colorScheme).opacity(colorScheme == .dark ? 0.72 : 0.9))
-				.frame(height: 0.5)
-				.allowsHitTesting(false)
-		}
-	}
-
-	private var emptyState: some View {
-		VStack(alignment: .leading, spacing: 6) {
-			Image(systemName: "person.crop.circle.badge.plus")
-				.font(PanelFont.emptyIcon)
-				.foregroundStyle(PanelPalette.secondaryText(colorScheme))
-			Text("No accounts in the local pool")
-				.font(PanelFont.emptyTitle)
-				.foregroundStyle(PanelPalette.primaryText(colorScheme))
-			Text("Add a ChatGPT login before switching the Codex auth file.")
-				.font(PanelFont.emptyBody)
-				.foregroundStyle(PanelPalette.secondaryText(colorScheme))
-				.fixedSize(horizontal: false, vertical: true)
-		}
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.padding(8)
-		.modernGlassSurface(cornerRadius: 9, depth: .row)
-	}
-
-	private var loadingState: some View {
-		HStack(spacing: 7) {
-			ProgressView()
-				.controlSize(.small)
-			Text("Loading accounts")
-				.font(PanelFont.emptyTitle)
-				.foregroundStyle(PanelPalette.primaryText(colorScheme))
-			Spacer()
-		}
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.padding(8)
-		.modernGlassSurface(cornerRadius: 9, depth: .row)
 	}
 
 	private var accountList: some View {
@@ -293,265 +147,4 @@ struct AccountPanelView: View {
 		}
 		.animation(PanelMotion.accountRemoval, value: accounts.map(\.id))
 	}
-
-	private var codexAuthLabel: String {
-		guard let auth = store.accountList?.codexAuth else {
-			return "No Codex auth"
-		}
-
-		if emailsHidden {
-			if let account = account(matching: auth.selector) {
-				return displayName(for: account)
-			}
-			let identity = auth.accountFingerprint.isEmpty ? auth.selector : auth.accountFingerprint
-			return AccountDisplay.alias(forIdentity: identity)
-		}
-
-		return AccountDisplay.compactEmail(auth.displayName)
-	}
-
-	private var decodexModeLabel: String {
-		guard let control = store.accountList?.control else {
-			return "Not loaded"
-		}
-
-		if let selector = control.accountSelector, selector.isEmpty == false {
-			if emailsHidden {
-				if let account = account(matching: selector) {
-					return "To \(displayName(for: account))"
-				}
-
-				return "To \(AccountDisplay.alias(forIdentity: selector))"
-			}
-
-			if selector.contains("@") {
-				return "To \(AccountDisplay.compactEmail(selector))"
-			}
-
-			return "To \(AccountDisplay.compactIdentity(selector))"
-		}
-
-		return control.mode
-	}
-
-	private var hasFixedSelection: Bool {
-		guard let selector = store.accountList?.control.accountSelector else {
-			return false
-		}
-
-		return selector.isEmpty == false
-	}
-
-	private var accountListContentHeight: CGFloat {
-		store.accounts.reduce(CGFloat(1)) { total, account in
-			total + accountRowHeight(for: account)
-		}
-	}
-
-	private var accountListViewportHeight: CGFloat {
-		min(accountListContentHeight, accountListAvailableHeight)
-	}
-
-	private var accountListNeedsScrolling: Bool {
-		accountListContentHeight > accountListAvailableHeight + 1
-	}
-
-	private var panelAnimationKey: AccountPanelAnimationKey {
-		AccountPanelAnimationKey(
-			accountIDs: store.accounts.map(\.id),
-			isInitialLoading: store.isInitialLoading,
-			hasAccounts: store.accounts.isEmpty == false,
-			hasTelemetry: telemetryMatrixIsVisible,
-			hasNotice: store.notice != nil,
-			hasUsageProbeError: store.accountList?.usageProbeError != nil,
-			hasFixedSelection: hasFixedSelection,
-			emailsHidden: emailsHidden,
-			needsScrolling: accountListNeedsScrolling
-		)
-	}
-
-	private var accountListAvailableHeight: CGFloat {
-		let visibleHeight = AccountPanelLayout.activeScreenVisibleHeight()
-		let availableHeight = visibleHeight - accountPanelChromeHeight
-		let minimumHeight = min(
-			AccountPanelLayout.minimumScrollableListHeight,
-			max(140, visibleHeight * 0.42)
-		)
-
-		return max(minimumHeight, availableHeight)
-	}
-
-	private var accountPanelChromeHeight: CGFloat {
-		var height = AccountPanelLayout.screenVerticalMargin
-			+ AccountPanelLayout.panelVerticalPadding
-			+ AccountPanelLayout.headerHeight
-			+ AccountPanelLayout.accountSummaryHeight
-			+ AccountPanelLayout.sectionSpacing
-
-		if telemetryMatrixIsVisible {
-			height += AccountPanelLayout.sectionSpacing + telemetryMatrixHeight
-		}
-		if store.notice != nil {
-			height += AccountPanelLayout.sectionSpacing + AccountPanelLayout.noticeHeight
-		}
-		if store.accountList?.usageProbeError != nil {
-			height += AccountPanelLayout.sectionSpacing + AccountPanelLayout.noticeHeight
-		}
-
-		return height
-	}
-
-	private var accountScrollProbe: some View {
-		GeometryReader { proxy in
-			Color.clear.preference(
-				key: AccountScrollOffsetPreferenceKey.self,
-				value: proxy.frame(in: .named(AccountPanelLayout.accountListScrollSpace)).minY
-			)
-		}
-	}
-
-	private var headerSubtitle: String {
-		let count = store.accounts.count
-		let accountLabel = "\(count) account\(count == 1 ? "" : "s")"
-		let routeLabel = hasFixedSelection ? "Routed" : "Balanced"
-		return "\(accountLabel) · \(routeLabel)"
-	}
-
-	private var emailsHidden: Bool {
-		accountPrivacy != AccountPrivacy.visibleValue
-	}
-
-	private var accountProfileAggregate: AccountProfileAggregate? {
-		AccountProfileAggregate.make(accounts: store.accounts)
-	}
-
-	private var telemetryMatrixIsVisible: Bool {
-		accountProfileAggregate != nil
-			|| store.accountList?.usageEstimate != nil
-	}
-
-	private var telemetryMatrixHeight: CGFloat {
-		var rows = [CGFloat]()
-		if accountProfileAggregate != nil {
-			rows.append(AccountPanelLayout.telemetryProfileHeight)
-		}
-		if let estimate = store.accountList?.usageEstimate {
-			rows.append(
-				estimate.accountEstimateCount < estimate.accountCount
-					? AccountPanelLayout.telemetryPoolMeasuredHeight
-					: AccountPanelLayout.telemetryPoolHeight
-			)
-		}
-		guard rows.isEmpty == false else {
-			return 0
-		}
-
-		return AccountPanelLayout.telemetryVerticalPadding
-			+ rows.reduce(0, +)
-			+ CGFloat(rows.count - 1) * AccountPanelLayout.telemetryRowSpacing
-	}
-
-	private func displayName(for account: CodexAccount) -> String {
-		if emailsHidden {
-			return AccountDisplay.aliases(for: store.accounts)[account.id]
-				?? account.panelDisplayName(emailsHidden: true)
-		}
-
-		return account.panelDisplayName(emailsHidden: false)
-	}
-
-	private var operatorCurrentLaneCards: [OperatorCurrentLaneCard] {
-		store.operatorPresentation?.currentLaneCards
-			?? store.operatorSnapshot?.presentation?.currentLaneCards
-			?? []
-	}
-
-	private func operatorCurrentLaneCards(for account: CodexAccount) -> [OperatorCurrentLaneCard] {
-		operatorCurrentLaneCards.filter { $0.isAssigned(to: account) }
-	}
-
-	private func accountRowHeight(for account: CodexAccount) -> CGFloat {
-		let base: CGFloat
-		if account.hasUsageWindowSummary {
-			base = 102
-		} else {
-			base = 48
-		}
-		let runSignal: CGFloat = operatorCurrentLaneCards(for: account).isEmpty ? 0 : 22
-		let profileSignal: CGFloat = account.hasProfileSummary
-			? (account.recentProfileDailyUsage.isEmpty ? 19 : 35)
-			: 0
-
-		return base + runSignal + profileSignal
-	}
-
-	private func account(matching selector: String) -> CodexAccount? {
-		store.accounts.first { account in
-			account.matchesSelector(selector)
-		}
-	}
-
-	private func requestLogout(_ account: CodexAccount) {
-		guard deletingLogoutAccountID == nil else {
-			return
-		}
-		if armedLogoutAccountID == account.id {
-			return
-		}
-
-		withAnimation(PanelMotion.inlineLayout) {
-			logoutErrorMessage = nil
-			armedLogoutAccountID = account.id
-		}
-	}
-
-	private func confirmLogout(_ account: CodexAccount) {
-		Task {
-			withAnimation(PanelMotion.accountRemoval) {
-				deletingLogoutAccountID = account.id
-				store.beginOptimisticLogoutRemoval(account)
-			}
-			logoutErrorMessage = nil
-			do {
-				try await store.logout(account)
-				deletingLogoutAccountID = nil
-				disarmLogout()
-			} catch {
-				withAnimation(PanelMotion.accountRemoval) {
-					store.cancelOptimisticLogoutRemoval(account)
-					deletingLogoutAccountID = nil
-					logoutErrorMessage = error.localizedDescription
-				}
-			}
-		}
-	}
-
-	private func disarmLogout() {
-		guard deletingLogoutAccountID == nil else {
-			return
-		}
-		withAnimation(PanelMotion.inlineLayout) {
-			logoutErrorMessage = nil
-			armedLogoutAccountID = nil
-		}
-	}
-
-	private func presentLogin(_ mode: AccountLoginSheetMode) {
-		loginWindowState.mode = mode
-		store.resetLoginSession()
-		NSApp.activate(ignoringOtherApps: true)
-		loginWindowState.isPresented = true
-	}
-}
-
-private struct AccountPanelAnimationKey: Equatable {
-	let accountIDs: [String]
-	let isInitialLoading: Bool
-	let hasAccounts: Bool
-	let hasTelemetry: Bool
-	let hasNotice: Bool
-	let hasUsageProbeError: Bool
-	let hasFixedSelection: Bool
-	let emailsHidden: Bool
-	let needsScrolling: Bool
 }
